navbar sin searchbar terminada
