+++ client/src/App.css
@@ -1,7 +1,7 @@
 #root {
   align-items: center;
   justify-content: center;
-  max-width: 1200px;
+  max-width: 1400px;
   margin: 0 auto;
   width: auto;
 }
