+++ client/src/components/NavBar/NavBar.module.css
@@ -1,5 +1,4 @@
 .logo {
     width: 10rem;
     object-fit: contain;
-    
 }
